fix: validate project names
