+++ bin/send-kubeconfig.js
@@ -1,10 +1,9 @@
 const { program } = require("commander");
 const consts = require("../lib/consts");
-const glob = require("glob");
 const nodemailer = require("nodemailer");
 const handlebars = require("handlebars");
-const { getTemplate, assertValidData } = require("../lib/templates");
-const { schemaGithubUsername } = require("../lib/schemas");
+const { getTemplate } = require("../lib/templates");
+const { schemaGithubUsername, assertValidData } = require("../lib/schemas");
 const { assertFile } = require("../lib/file");
 const { getKubeconfigFileForUsername } = require("../lib/consts");
 const { readClusterInfo } = require("../lib/cluster-info");
