+++ lib/schemas.js
@@ -38,7 +38,16 @@ const schemaTerraformOutput = Joi.object()
   })
   .options({ allowUnknown: true });
 
+const assertValidData = (data, schema) => {
+  const validationResult = schema.validate(data);
+  if (validationResult.error) {
+    console.error(validationResult.error);
+    process.exit(1);
+  }
+};
+
 module.exports = {
+  assertValidData,
   schemaProjectName,
   schemaProjectStage,
   schemaBuckets,
