+++ lib/templates.js
@@ -11,16 +11,7 @@ const getTemplate = (handlebars, category, file) => {
 const renderToFile = (template, data, dest) =>
   fs.writeFileSync(dest, template(data));
 
-const assertValidData = (data, schema) => {
-  const validationResult = schema.validate(data);
-  if (validationResult.error) {
-    console.error(validationResult.error);
-    process.exit(1);
-  }
-};
-
 module.exports = {
   getTemplate,
   renderToFile,
-  assertValidData,
 };
